messy write support
A few things were disabled to limit the extent of the changes:
- streamed blocks
- external blocks
- update
- resolve_and_inline

These changes are also a work-in-progress, more of an attempt
to see if the updated strategy will work.
