+++ asdf/_block/__init__.py
@@ -1,5 +1,5 @@
 from .block import Block, UnloadedBlock
-from .manager import BlockManager
+from .old_manager import BlockManager
 from .util import calculate_updated_layout
 
 __all__ = ["Block", "UnloadedBlock", "BlockManager", "calculate_updated_layout"]
