+++ asdf/_block/writer.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 from asdf import constants
 
 from . import io as bio
@@ -15,6 +17,10 @@ class WriteBlock:
             return self._data()
         return self._data
 
+    @property
+    def data_bytes(self):
+        return np.ndarray(-1, np.uint8, self.data.ravel(order="K").data)
+
 
 def write_blocks(fd, blocks, padding=False, streamed_block=None, write_index=True):
     offsets = []
@@ -25,7 +31,7 @@ def write_blocks(fd, blocks, padding=False, streamed_block=None, write_index=Tru
         headers.append(
             bio.write_block(
                 fd,
-                blk.data,
+                blk.data_bytes,
                 compression_kwargs=blk.compression_kwargs,
                 padding=padding,
                 compression=blk.compression,
@@ -34,7 +40,7 @@ def write_blocks(fd, blocks, padding=False, streamed_block=None, write_index=Tru
     if streamed_block is not None:
         offsets.append(fd.tell())
         fd.write(constants.BLOCK_MAGIC)
-        headers.append(bio.write_block(fd, streamed_block.data, stream=True))
+        headers.append(bio.write_block(fd, streamed_block.data_bytes, stream=True))
     elif len(blocks) and write_index:
         bio.write_block_index(fd, offsets)
     return offsets, headers
