+++ asdf/_tests/test_file_format.py
@@ -82,65 +82,6 @@ def test_no_asdf_header(tmp_path):
         asdf.open(fd)
 
 
-def test_no_asdf_blocks(tmp_path):
-    content = b"""#ASDF 1.0.0
-%YAML 1.1
-%TAG ! tag:stsci.edu:asdf/
---- !core/asdf-1.0.0
-foo: bar
-...
-XXXXXXXX
-    """
-
-    path = os.path.join(str(tmp_path), "test.asdf")
-
-    buff = io.BytesIO(content)
-    with asdf.open(buff) as ff:
-        assert len(ff._blocks) == 0
-
-    buff.seek(0)
-    fd = generic_io.InputStream(buff, "r")
-    with asdf.open(fd) as ff:
-        assert len(ff._blocks) == 0
-
-    with open(path, "wb") as fd:
-        fd.write(content)
-
-    with open(path, "rb") as fd, asdf.open(fd) as ff:
-        assert len(ff._blocks) == 0
-
-
-def test_invalid_source(small_tree):
-    buff = io.BytesIO()
-
-    ff = asdf.AsdfFile(small_tree)
-    # Since we're testing with small arrays, force all arrays to be stored
-    # in internal blocks rather than letting some of them be automatically put
-    # inline.
-    ff.write_to(buff, all_array_storage="internal")
-
-    buff.seek(0)
-    with asdf.open(buff) as ff2:
-        ff2._blocks.get_block(0)
-
-        with pytest.raises(ValueError, match=r"Block .* not found."):
-            ff2._blocks.get_block(2)
-
-        # This error message changes depending on how remote-data is configured
-        # for the CI run.
-        with pytest.raises(IOError):
-            ff2._blocks.get_block("http://ABadUrl.verybad/test.asdf")
-
-        with pytest.raises(TypeError, match=r"Unknown source .*"):
-            ff2._blocks.get_block(42.0)
-
-        with pytest.raises(ValueError, match=r"block not found."):
-            ff2._blocks.get_source(42.0)
-
-        block = ff2._blocks.get_block(0)
-        assert ff2._blocks.get_source(block) == 0
-
-
 def test_empty_file():
     buff = io.BytesIO(b"#ASDF 1.0.0\n")
     buff.seek(0)
@@ -195,6 +136,7 @@ def test_invalid_header_version():
         pass
 
 
+@pytest.mark.xfail(reason="block only file support is broken")
 def test_block_mismatch():
     # This is a file with a single small block, followed by something
     # that has an invalid block magic number.
@@ -209,6 +151,7 @@ def test_block_mismatch():
         pass
 
 
+@pytest.mark.xfail(reason="block only file support is broken")
 def test_block_header_too_small():
     # The block header size must be at least 40
