+++ asdf/_tests/test_yaml.py
@@ -203,8 +203,7 @@ def test_explicit_tags():
 %YAML 1.1
 --- !<tag:stsci.edu:asdf/core/asdf-1.1.0>
 foo: !<tag:stsci.edu:asdf/core/ndarray-1.0.0> [1, 2, 3]
-...
-    """
+..."""
 
     # Check that fully qualified explicit tags work
     buff = helpers.yaml_to_asdf(yaml, yaml_headers=False)
